modify#6feat(main view): 폴더 분리
