--- frontEnd/public/js/init.js
@@ -1,3 +1,0 @@
-import { eventToBtns } from "./main/component/button.js ";
-
-eventToBtns();
