+++ frontEnd/public/views/mypage/chart.js
@@ -1,4 +1,4 @@
-import { navBarView } from "../navBarView";
+import { navBarView } from "../../navbar/view";
 
 const chartView = () => {
     const $chartContainer = document.createElement("section");
