+++ frontEnd/public/views/mypage/comparison.js
@@ -1,4 +1,4 @@
-import { navBarView } from "../navBarView";
+import { navBarView } from "../../navbar/view";
 
 const comparisonView = () => {
     const $comparionContainer = document.createElement("section");
